Tidied up the integer conversions between doubles and long in the organs expiry time uses. The hours until expiry in the Organs enum takes a double in case there is something less than one hour, but once it is converted to seconds it is cast to a long because the java.time methods use longs
#story[1017]

+++ commons/src/main/java/odms/commons/model/_enum/Organs.java
@@ -36,13 +36,13 @@ public enum Organs {
     private String organName;
     private int dbValue;
     private static final int HOURS_TO_SECONDS = 3600;
-    private int storageSeconds;
+    private long storageSeconds;
 
 
     Organs(String organName, int dbValue, double storageSeconds) {
         this.organName = organName;
         this.dbValue = dbValue;
-        this.storageSeconds = (int) (storageSeconds * HOURS_TO_SECONDS);
+        this.storageSeconds = (long)(storageSeconds * HOURS_TO_SECONDS);
     }
 
     @Override
@@ -54,7 +54,7 @@ public enum Organs {
         return dbValue;
     }
 
-    public int getStorageSeconds() {
+    public long getStorageSeconds() {
         return storageSeconds;
     }
 }
